Moving browsers to launcher
