+++ packages/frontend-shared/cypress/e2e/support/e2eSupport.ts
@@ -1,5 +1,5 @@
 import '@testing-library/cypress/add-commands'
-import { browsers } from '@packages/types/src/browser'
+import { browsers } from '@packages/launcher/lib/browsers'
 import { installCustomPercyCommand } from '@packages/ui-components/cypress/support/customPercyCommand'
 import { configure } from '@testing-library/cypress'
 import i18n from '../../../src/locales/en-US.json'
