+++ packages/launcher/test/unit/windows_spec.ts
@@ -138,8 +138,6 @@ describe('windows browser detection', () => {
           version: '100',
           majorVersion: '100',
           path: win10Path,
-          unsupportedVersion: false,
-          warning: undefined,
         }),
       )
     })
